+++ server/src/test/passage-text-parsers/sugarcube/sugarcube-diagnostics.test.ts
@@ -11,7 +11,6 @@ import { TextDocument } from "vscode-languageserver-textdocument";
 
 import { DiagnosticCodes } from "../../../diagnostics";
 import { Index } from "../../../project-index";
-import { defaultDiagnosticsOptions } from "../../../server-options";
 import { OSugarCubeSymbolKind } from "../../../passage-text-parsers/sugarcube/types";
 
 import * as uut from "../../../passage-text-parsers/sugarcube";
@@ -35,14 +34,9 @@ describe("SugarCube Diagnostics", () => {
                     kind: OSugarCubeSymbolKind.Variable,
                 },
             ]);
-            const diagnosticOptions = defaultDiagnosticsOptions;
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.eql([
                 Diagnostic.create(
@@ -75,14 +69,9 @@ describe("SugarCube Diagnostics", () => {
             index.setDisabledDiagnosticRanges("fake-uri", {
                 [DiagnosticCodes.VariableNeverSet]: [Range.create(1, 1, 6, 1)],
             });
-            const diagnosticOptions = defaultDiagnosticsOptions;
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.be.empty;
         });
@@ -113,14 +102,9 @@ describe("SugarCube Diagnostics", () => {
                     kind: OSugarCubeSymbolKind.VariableSet,
                 },
             ]);
-            const diagnosticOptions = defaultDiagnosticsOptions;
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.be.empty;
         });
@@ -142,14 +126,9 @@ describe("SugarCube Diagnostics", () => {
                     kind: OSugarCubeSymbolKind.Variable,
                 },
             ]);
-            const diagnosticOptions = defaultDiagnosticsOptions;
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.be.empty;
         });
@@ -171,14 +150,9 @@ describe("SugarCube Diagnostics", () => {
                     kind: OSugarCubeSymbolKind.Property,
                 },
             ]);
-            const diagnosticOptions = defaultDiagnosticsOptions;
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.be.empty;
         });
@@ -200,14 +174,9 @@ describe("SugarCube Diagnostics", () => {
                     kind: OSugarCubeSymbolKind.Property,
                 },
             ]);
-            const diagnosticOptions = defaultDiagnosticsOptions;
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.be.empty;
         });
@@ -229,14 +198,9 @@ describe("SugarCube Diagnostics", () => {
                     kind: OSugarCubeSymbolKind.Property,
                 },
             ]);
-            const diagnosticOptions = defaultDiagnosticsOptions;
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.be.empty;
         });
@@ -260,15 +224,9 @@ describe("SugarCube Diagnostics", () => {
                     kind: OSugarCubeSymbolKind.UnknownMacro,
                 },
             ]);
-            const diagnosticOptions = defaultDiagnosticsOptions;
-            diagnosticOptions.warnings.unknownMacro = true;
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.eql([
                 Diagnostic.create(
@@ -303,15 +261,9 @@ describe("SugarCube Diagnostics", () => {
                     Range.create(1, 1, 6, 1),
                 ],
             });
-            const diagnosticOptions = defaultDiagnosticsOptions;
-            diagnosticOptions.warnings.unknownMacro = true;
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.be.empty;
         });
@@ -343,45 +295,9 @@ describe("SugarCube Diagnostics", () => {
                     kind: OSugarCubeSymbolKind.KnownMacro,
                 },
             ]);
-            const diagnosticOptions = defaultDiagnosticsOptions;
-            diagnosticOptions.warnings.unknownMacro = true;
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
-
-            expect(results).to.be.empty;
-        });
-
-        it("should not warn on an unrecognized macro if that warning is disabled", () => {
-            const doc = TextDocument.create(
-                "fake-uri",
-                "",
-                0,
-                "Let's try <<this>>",
-            );
-            const index = new Index();
-            index.setReferences("fake-uri", [
-                {
-                    contents: "this",
-                    locations: [
-                        Location.create("fake-uri", Range.create(1, 2, 3, 4)),
-                    ],
-                    kind: OSugarCubeSymbolKind.UnknownMacro,
-                },
-            ]);
-            const diagnosticOptions = defaultDiagnosticsOptions;
-            diagnosticOptions.warnings.unknownMacro = false;
-            const parser = uut.getSugarCubeParser(undefined);
-
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                diagnosticOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.be.empty;
         });
@@ -401,11 +317,7 @@ describe("SugarCube Diagnostics", () => {
             ]);
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                defaultDiagnosticsOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.eql([
                 Diagnostic.create(
@@ -438,11 +350,7 @@ describe("SugarCube Diagnostics", () => {
             });
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                defaultDiagnosticsOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.be.empty;
         });
@@ -472,11 +380,7 @@ describe("SugarCube Diagnostics", () => {
             ]);
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                defaultDiagnosticsOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.eql([
                 Diagnostic.create(
@@ -528,11 +432,7 @@ describe("SugarCube Diagnostics", () => {
             });
             const parser = uut.getSugarCubeParser(undefined);
 
-            const results = parser?.generateDiagnostics(
-                doc,
-                index,
-                defaultDiagnosticsOptions,
-            );
+            const results = parser?.generateDiagnostics(doc, index);
 
             expect(results).to.be.empty;
         });